+++ src/Candoumbe.Pipelines/Components/ICompile.cs
@@ -1,10 +1,10 @@
-﻿using System;
+using System;
 using Candoumbe.Pipelines.Components.Formatting;
-using Nuke.Common;
-using Nuke.Common.Tooling;
-using Nuke.Common.Tools.DotNet;
-using Nuke.Common.Utilities.Collections;
-using static Nuke.Common.Tools.DotNet.DotNetTasks;
+using Fallout.Common;
+using Fallout.Common.Tooling;
+using Fallout.Common.Tools.DotNet;
+using Fallout.Common.Utilities.Collections;
+using static Fallout.Common.Tools.DotNet.DotNetTasks;
 using static Serilog.Log;
 
 namespace Candoumbe.Pipelines.Components;