+++ CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### 💥 Breaking changes
+- Migrated from [Nuke](https://github.com/nuke-build/nuke) to [Fallout](https://github.com/fallout-build/fallout)
+
 ### 🚀 New features
+- Added support for [Fallout](https://github.com/Fallout-build/Fallout) as build engine (migration from [Nuke](https://github.com/nuke-build/nuke))
 - Added `net10.0` support
 - Added support for autoclosing issues when finishing a hotfix via the `--issues` parameter ([#177](https://github.com/candoumbe/pipelines/issues/177))
   - GitHub: Closes issues using GitHub API